+++ plotnine/scales/scale_xy.py
@@ -24,6 +24,11 @@ class scale_position_discrete(scale_discrete):
     Parameters
     ----------
     {superclass_parameters}
+    limits : array_like, optional
+        Limits of the scale. For discrete scale, these are
+        the categories (unique values) of the variable.
+        For scales that deal with categoricals, these may
+        be a subset or superset of the categories.
     """
     # All positions have no guide
     guide = None
@@ -77,12 +82,17 @@ class scale_position_discrete(scale_discrete):
     def limits(self):
         if self.is_empty():
             return (0, 1)
-
-        if self._limits is not None:
+        elif self._limits is not None and not callable(self._limits):
             return self._limits
-        elif self.range.range:
+        elif self._limits is None:
             # discrete range
             return self.range.range
+        elif callable(self._limits):
+            limits = self._limits(self.range.range)
+            # Functions that return iterators e.g. reversed
+            if iter(limits) is limits:
+                limits = list(limits)
+            return limits
         else:
             raise PlotnineError(
                 'Lost, do not know what the limits are.')